app.py
import streamlit as st
import cv2
import numpy as np
from PIL import Image
import tempfile
import os

st.set_page_config(
    page_title="FakeShield - Deepfake Detection",
    page_icon="🛡️",
    layout="wide"
)

def main():
    st.title("🛡️ FakeShield")
    st.subheader("AI-powered Deepfake Video & Voice Detection Tool")
    
    tab1, tab2 = st.tabs(["Video Detection", "Audio Detection"])
    
    with tab1:
        st.header("Video Deepfake Detection")
        uploaded_video = st.file_uploader(
            "Choose a video file", 
            type=['mp4', 'avi', 'mov', 'mkv'],
            key="video_uploader"
        )
        
        if uploaded_video is not None:
            # Save uploaded file temporarily
            with tempfile.NamedTemporaryFile(delete=False, suffix='.mp4') as tmp_file:
                tmp_file.write(uploaded_video.read())
                tmp_file_path = tmp_file.name
            
            st.video(uploaded_video)
            
            if st.button("Analyze Video", key="analyze_video"):
                with st.spinner("Analyzing video for deepfake indicators..."):
                    # Placeholder for actual deepfake detection logic
                    result = analyze_video_deepfake(tmp_file_path)
                    
                    col1, col2 = st.columns(2)
                    with col1:
                        st.metric("Authenticity Score", f"{result['score']:.1f}%")
                    with col2:
                        st.metric("Confidence", f"{result['confidence']:.1f}%")
                    
                    if result['score'] > 70:
                        st.success("✅ Video appears to be authentic")
                    elif result['score'] > 30:
                        st.warning("⚠️ Video shows some suspicious patterns")
                    else:
                        st.error("🚨 High probability of deepfake detection")
            
            # Clean up temp file
            if os.path.exists(tmp_file_path):
                os.unlink(tmp_file_path)
    
    with tab2:
        st.header("Audio Deepfake Detection")
        uploaded_audio = st.file_uploader(
            "Choose an audio file", 
            type=['wav', 'mp3', 'ogg', 'm4a'],
            key="audio_uploader"
        )
        
        if uploaded_audio is not None:
            st.audio(uploaded_audio)
            
            if st.button("Analyze Audio", key="analyze_audio"):
                with st.spinner("Analyzing audio for voice synthesis indicators..."):
                    # Placeholder for actual voice deepfake detection logic
                    result = analyze_audio_deepfake(uploaded_audio)
                    
                    col1, col2 = st.columns(2)
                    with col1:
                        st.metric("Authenticity Score", f"{result['score']:.1f}%")
                    with col2:
                        st.metric("Confidence", f"{result['confidence']:.1f}%")
                    
                    if result['score'] > 70:
                        st.success("✅ Audio appears to be authentic")
                    elif result['score'] > 30:
                        st.warning("⚠️ Audio shows some suspicious patterns")
                    else:
                        st.error("🚨 High probability of voice synthesis detection")

def analyze_video_deepfake(video_path):
    """
    Placeholder function for video deepfake detection.
    In a real implementation, this would use advanced ML models.
    """
    # Simulate analysis with random but realistic results
    import random
    score = random.uniform(15, 95)
    confidence = random.uniform(60, 95)
    
    return {
        'score': score,
        'confidence': confidence,
        'details': 'Analysis complete'
    }

def analyze_audio_deepfake(audio_file):
    """
    Placeholder function for audio deepfake detection.
    In a real implementation, this would analyze audio patterns.
    """
    # Simulate analysis with random but realistic results
    import random
    score = random.uniform(20, 90)
    confidence = random.uniform(65, 90)
    
    return {
        'score': score,
        'confidence': confidence,
        'details': 'Audio analysis complete'
    }

if __name__ == "__main__":
    main()
